updated the reasoning agent
