+++ package.json
@@ -31,7 +31,7 @@
     "@langchain/anthropic": "^0.2.3",
     "@langchain/azure-openai": "^0.0.11",
     "@langchain/community": "^0.3.15",
-    "@langchain/core": "^0.3.30",
+    "@langchain/core": "^0.3.40",
     "@langchain/openai": "^0.0.25",
     "@xenova/transformers": "^2.17.1",
     "axios": "^1.6.8",
@@ -43,9 +43,10 @@
     "drizzle-orm": "^0.31.2",
     "express": "^4.19.2",
     "html-to-text": "^9.0.5",
-    "langchain": "^0.3.13",
+    "langchain": "^0.3.15",
     "multer": "^1.4.5-lts.1",
     "next": "^15.1.6",
+    "openai": "^4.85.1",
     "pdf-parse": "^1.1.1",
     "postgres": "^3.4.5",
     "razorpay": "^2.9.5",
