+++ src/websocket/messageHandler.ts
@@ -2,8 +2,8 @@ import { EventEmitter, WebSocket } from 'ws';
 import { BaseMessage, AIMessage, HumanMessage } from '@langchain/core/messages';
 import handleWebSearch from '../agents/webSearchAgent';
 import handleAcademicSearch from '../agents/academicSearchAgent';
-// import handleWritingAssistant from '../agents/writingAssistant';
-import handleReasoningAgent from '../agents/writingAssistant';
+import handleWritingAssistant from '../agents/writingAssistant';
+import handleReasoningAgent from '../agents/reasoningAssistant';
 import handleWolframAlphaSearch from '../agents/wolframAlphaSearchAgent';
 import handleYoutubeSearch from '../agents/youtubeSearchAgent';
 import handleRedditSearch from '../agents/redditSearchAgent';
@@ -33,11 +33,11 @@ type WSMessage = {
 export const searchHandlers = {
   webSearch: handleWebSearch,
   academicSearch: handleAcademicSearch,
-  writingAssistant: handleReasoningAgent,
-  // handleWritingAssistant,
+  writingAssistant:handleWritingAssistant,
   wolframAlphaSearch: handleWolframAlphaSearch,
   youtubeSearch: handleYoutubeSearch,
   redditSearch: handleRedditSearch,
+  reasoning:handleReasoningAgent
 };
 
 const handleEmitterEvents = (
