--- uploads/b8cdc74f9e0d4e4919aafc4b9d90e768-extracted.json
@@ -1,1 +1,0 @@
-{"title":"Anshu_cover _letter.pdf","contents":["Anshu Kumar Bishwas\nForbesganj, Araria, Bihar –854318\n+91 7667411679 –anshubishwas0@gmail.com\nDearHiringManager,\nIamapassionateandmotivatedFull-StackWebDeveloperwithastrongfoundationinawiderangeofwebtechnologies.Myexperienceincludes\ndevelopinganddeployingcomprehensivesystems,aswellascompletingafrontenddevelopmentinternshipthathasfurthersharpenedmyskillsand\ndeepenedmyunderstandingofthedevelopmentlifecycle.\nKeySkills:\n•ProgrammingLanguages:C/C++,Java,HTML/CSS,JavaScript,PHP,SQL","KeySkills:\n•ProgrammingLanguages:C/C++,Java,HTML/CSS,JavaScript,PHP,SQL\n•Tools&Frameworks:Node.js,React.js,MongoDB,Git/GitHub,Dialogflow\nIameagertocontributemytechnicalexpertiseandenthusiasmforwebdevelopmenttoyourteam.Iamconfidentthatmycombinationofskills,practical\nexperience,andpassionwillallowmetomakeameaningfulimpactinyourorganization.\nThankyouforconsideringmyapplication.\nSincerely,\nAnshuKumarBishwas"]}
